[Generator] Create png canvas (#2)

+++ gutenku-api/bin/fetch_book.py
@@ -1,22 +1,30 @@
 import os
+import sys
 import requests
-from dotenv import load_dotenv
 
-load_dotenv(os.path.join(os.path.dirname(os.path.dirname(__file__)), '.env'))
-
-book_id = os.environ.get('BOOK_ID')
+book_id = int(sys.argv[1])
 url = f"https://www.gutenberg.org/cache/epub/{book_id}/pg{book_id}.txt"
 
-# Make the request to the URL and get the response
-response = requests.get(url)
+cache_directory = "../.cache"
 
-# Check that the response is OK
-if response.status_code == 200:
-    cache_directory = "../.cache"
+if not os.path.exists(cache_directory):
+    os.makedirs(cache_directory)
 
-    with open(f"{cache_directory}/book_{book_id}.txt", "w", encoding="utf-8") as f:
-        f.write(response.text)
-        
-    print(f"The book {book_id} has been successfully fetched")
-else:
-    print("The request failed with status code {}".format(response.status_code))
+# Define the path to the text file containing the ebook
+file_path = f"{cache_directory}/book_{book_id}.txt"
+
+if not os.path.exists(file_path):
+    # Make the request to the URL and get the response
+    response = requests.get(url)
+
+    # Check that the response is OK
+    if response.status_code == 200:
+        cache_directory = "../.cache"
+
+        with open(f"{cache_directory}/book_{book_id}.txt", "w", encoding="utf-8") as f:
+            f.write(response.text)
+            
+        print(f"The book \033[1;32m{book_id}\033[0m has been successfully fetched")
+    else:
+        print("The request failed with status code {}".format(response.status_code))
+
